requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
flask-login>=0.6.3
gunicorn>=22.0.0
eventlet>=0.33.3
werkzeug>=3.0.0
flask-wtf>=1.2.1
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
pytest
pytest-flask
pytest-mock
google-generativeai>=0.3.0
flask-compress
flask-limiter